+++ 5_evaluate_baseline.sh
@@ -9,10 +9,10 @@ cd $SCRIPT_DIR
 
 echo
 echo '*** Evaluate code-llama base model ***'
-echo uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc --config config_initial.json base test --save_results_file evaluation/code_llama_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
-uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc --config config_initial.json base test --save_results_file evaluation/code_llama_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
-echo "uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc-data-to-md evaluation/code_llama_base_results.json > evaluation/code_llama_base_results.md"
-uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc-data-to-md evaluation/code_llama_base_results.json > evaluation/code_llama_base_results.md
+echo uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc --config config_initial.json base test --save_results_file evaluation/code_llama_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
+uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc --config config_initial.json base test --save_results_file evaluation/code_llama_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
+echo "uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc-data-to-md evaluation/code_llama_base_results.json > evaluation/code_llama_base_results.md"
+uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc-data-to-md evaluation/code_llama_base_results.json > evaluation/code_llama_base_results.md
 
 echo
 echo '*** Evaluate OpenAI gpt-4o base model ***'
@@ -23,9 +23,9 @@ if [ -z "$OPENAI_API_KEY" ]; then
     exit 1
 else
     echo "OPENAI_API_KEY is set."
-    echo uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc --config config_initial.json openai test --model gpt-4o --save_results_file evaluation/openai_gpt-4o_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
-    uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc --config config_initial.json openai test --model gpt-4o --save_results_file evaluation/openai_gpt-4o_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
-    echo "uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc-data-to-md evaluation/openai_gpt-4o_base_results.json > evaluation/openai_gpt-4o_base_results.md"
-    uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.4.1 chathpc-data-to-md evaluation/openai_gpt-4o_base_results.json > evaluation/openai_gpt-4o_base_results.md
+    echo uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc --config config_initial.json openai test --model gpt-4o --save_results_file evaluation/openai_gpt-4o_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
+    uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc --config config_initial.json openai test --model gpt-4o --save_results_file evaluation/openai_gpt-4o_base_results.json C2_Kokkos_Dataset/kokkos_testing.yaml
+    echo "uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc-data-to-md evaluation/openai_gpt-4o_base_results.json > evaluation/openai_gpt-4o_base_results.md"
+    uv run --project C1_ChatHPC_Lib/ChatHPC-app-v25.7.1 chathpc-data-to-md evaluation/openai_gpt-4o_base_results.json > evaluation/openai_gpt-4o_base_results.md
 fi