+++ C1_ChatHPC_Lib/ChatHPC-app-v25.7.1/tests/test_app.py
@@ -8,7 +8,7 @@ def test_chat_prompt_json():
     """Test basic prompt function from reading JSON config."""
     app = App.from_json("tests/files/config.json")
 
-    expected = "You are a powerful LLM model for Kokkos called ChatKokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Context:\nContext\n\n### Question:\nQuestion\n\n### Answer:\n"
+    expected = "You are a powerful LLM model for Kokkos called ChatHPC for Kokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Context:\nContext\n\n### Question:\nQuestion\n\n### Answer:\n"
     result = app.chat_prompt(question="Question", context="Context")
     assert result == expected, "Prompt is not as expected."
 
@@ -17,7 +17,7 @@ def test_chat_prompt_json_no_context():
     """Test without context"""
     app = App.from_json("tests/files/config.json")
 
-    expected = "You are a powerful LLM model for Kokkos called ChatKokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Question:\nQuestion\n\n### Answer:\n"
+    expected = "You are a powerful LLM model for Kokkos called ChatHPC for Kokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Question:\nQuestion\n\n### Answer:\n"
     result = app.chat_prompt(question="Question")
     assert result == expected, "Prompt is not as expected."
 
@@ -27,7 +27,7 @@ def test_training_prompt_json():
     app = App.from_json(
         "tests/files/config.json",
     )
-    expected = "You are a powerful LLM model for Kokkos called ChatKokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Context:\nContext\n\n### Question:\nQuestion\n\n### Answer:\nAnswer\n\n"
+    expected = "You are a powerful LLM model for Kokkos called ChatHPC for Kokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Context:\nContext\n\n### Question:\nQuestion\n\n### Answer:\nAnswer\n\n"
 
     result = app.training_prompt(user="Question", context="Context", assistant="Answer")
     assert result == expected, "Prompt is not as expected."
@@ -43,7 +43,7 @@ def test_training_prompt_json2():
         "context": "Introduction to Kokkos programming model",
         "answer": "```cpp\n#include <Kokkos_Core.hpp>\nint main( int argc, char* argv[] ) {\n  int M = 10;\n  Kokkos::initialize( argc, argv );{\n    Kokkos::View<float *> X(M);\n    Kokkos::View<float **> Y(M, M);\n    Kokkos::kokkos_free<>(X);\n    Kokkos::kokkos_free<>(Y);\n  }\n  Kokkos::finalize();\n  return 0;\n}\n```",
     }
-    expected = "You are a powerful LLM model for Kokkos called ChatKokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Context:\nIntroduction to Kokkos programming model\n\n### Question:\nCan you give me an example of a Kokkos view?\n\n### Answer:\n```cpp\n#include <Kokkos_Core.hpp>\nint main( int argc, char* argv[] ) {\n  int M = 10;\n  Kokkos::initialize( argc, argv );{\n    Kokkos::View<float *> X(M);\n    Kokkos::View<float **> Y(M, M);\n    Kokkos::kokkos_free<>(X);\n    Kokkos::kokkos_free<>(Y);\n  }\n  Kokkos::finalize();\n  return 0;\n}\n```\n\n"
+    expected = "You are a powerful LLM model for Kokkos called ChatHPC for Kokkos created by ORNL. Your job is to answer questions about the Kokkos programming model. You are given a question and context regarding the Kokkos programming model.\n\nYou must output the answer the question.\n\n### Context:\nIntroduction to Kokkos programming model\n\n### Question:\nCan you give me an example of a Kokkos view?\n\n### Answer:\n```cpp\n#include <Kokkos_Core.hpp>\nint main( int argc, char* argv[] ) {\n  int M = 10;\n  Kokkos::initialize( argc, argv );{\n    Kokkos::View<float *> X(M);\n    Kokkos::View<float **> Y(M, M);\n    Kokkos::kokkos_free<>(X);\n    Kokkos::kokkos_free<>(Y);\n  }\n  Kokkos::finalize();\n  return 0;\n}\n```\n\n"
 
     result = app.training_prompt(**inputd)
     assert result == expected, "Prompt is not as expected."